backend/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
python-dotenv==1.2.1
pydantic[email]==2.12.5

# Database
SQLAlchemy==2.0.48
alembic==1.18.4
psycopg2-binary==2.9.11

# Auth
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
cryptography==46.0.5

# Data
httpx
pandas
yfinance
